Day09/auth_system_V02/requirements.txt
Flask
mysql-connector-python
Werkzeug
PyJWT
